Implemented onClick for EventFragment
1. Implemented deletion of appointment via onItemLongClick method
2. Modified EventActivity such that quantity selected under everyNum and remindNum will be reflected in everyBox and remindBox

+++ app/src/main/java/mooncakemonster/orbitalcalendar/database/AppointmentController.java
@@ -38,7 +38,7 @@ public class AppointmentController
         dbHelper.close();
     }
 
-    public Appointment createAppointment(String event, String startproperdate, long startdate, long enddate, String location, String notes, int remind) {
+    public Appointment createAppointment(String event, String startproperdate, long startdate, long enddate, String location, String notes, long remind) {
         ContentValues values = new ContentValues();
         //Insert key-value in ContentValues
         values.put(DatabaseHelper.EVENT, event);
